+++ CHANGES.rst
@@ -1,5 +1,11 @@
 .. currentmodule:: click
 
+Version 8.4.2
+-------------
+
+Unreleased
+
+
 Version 8.4.1
 -------------